cc voznesenskym penguinwu EikanWang jgong5 Guobing-Chen XiaobingSuper zhuhaozhe blzheng wenzhe-nrv jiayisunx peterbell10 ipiszy yf225 chenyang78 kadeng muchulee8 aakhundov ColinPeppler

[ghstack-poisoned]

+++ torch/csrc/dynamo/guards.cpp
@@ -1,4 +1,5 @@
 #define PY_SSIZE_T_CLEAN
+#include <ATen/EmptyTensor.h>
 #include <c10/util/flat_hash_map.h>
 #include <torch/csrc/autograd/grad_mode.h>
 #include <torch/csrc/dynamo/guards.h>
@@ -622,8 +623,8 @@ static PyObject* _empty_strided_cpu(PyObject* dummy, PyObject* args) {
   // at::empty_strided is surprising slow.  This is a lower-overhead
   // version that saves ~2us on every allocation.
   HANDLE_TH_ERRORS;
-  at::SmallVector<ssize_t, 8> sizes;
-  at::SmallVector<ssize_t, 8> strides;
+  at::SmallVector<int64_t, 8> sizes;
+  at::SmallVector<int64_t, 8> strides;
   at::ScalarType dtype;
   _parse_empty_strided_args(args, sizes, strides, dtype);
   return THPVariable_Wrap(at::detail::empty_strided_cpu(sizes, strides, dtype));
@@ -634,8 +635,8 @@ static PyObject* _empty_strided_cuda(PyObject* dummy, PyObject* args) {
   // at::empty_strided is surprising slow.  This is lower-overhead.
   HANDLE_TH_ERRORS;
 #ifdef USE_CUDA
-  at::SmallVector<ssize_t, 8> sizes;
-  at::SmallVector<ssize_t, 8> strides;
+  at::SmallVector<int64_t, 8> sizes;
+  at::SmallVector<int64_t, 8> strides;
   at::ScalarType dtype;
   _parse_empty_strided_args(args, sizes, strides, dtype);
   return THPVariable_Wrap(at::detail::empty_strided_cuda(
